requirements.txt
fastapi==0.100.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.6
uvicorn==0.23.1
pydantic>=1.0,<3.0
python-dotenv
alembic==1.11.1
httpx==0.24.1
pytest==7.4.0
requests==2.31.0
pytest_expect==1.1.0
six==1.16.0
fastapi-cache2[redis]==0.2.1
aioredis==1.3.1
redis==4.6.0
pre-commit==3.3.3
types-redis==4.6.0.3
mypy==1.4.1
fastapi_sqlalchemy==0.2.1
pytest-asyncio==0.21.1
httpx==0.24.1
asyncpg==0.28.0
celery==5.3.1
flower==2.0.0
rabbitmq==0.2.0
cachetools==5.3.1
pandas==2.0.3
